+++ piccolo_admin/endpoints.py
@@ -5,6 +5,7 @@ Creates a basic wrapper around a Piccolo model, turning it into an ASGI app.
 from __future__ import annotations
 
 import inspect
+import io
 import itertools
 import json
 import logging
@@ -14,6 +15,7 @@ from dataclasses import dataclass
 from datetime import timedelta
 from functools import partial
 
+import typing_extensions
 from fastapi import FastAPI, File, Form, UploadFile
 from piccolo.apps.user.tables import BaseUser
 from piccolo.columns.base import Column
@@ -35,6 +37,8 @@ from piccolo_api.csrf.middleware import CSRFMiddleware
 from piccolo_api.fastapi.endpoints import FastAPIKwargs, FastAPIWrapper
 from piccolo_api.media.base import MediaStorage
 from piccolo_api.media.local import LocalMediaStorage
+from piccolo_api.mfa.endpoints import mfa_setup
+from piccolo_api.mfa.provider import MFAProvider
 from piccolo_api.openapi.endpoints import swagger_ui
 from piccolo_api.rate_limiting.middleware import (
     InMemoryLimitProvider,
@@ -49,7 +53,7 @@ from starlette.middleware import Middleware
 from starlette.middleware.authentication import AuthenticationMiddleware
 from starlette.middleware.exceptions import HTTPException
 from starlette.requests import Request
-from starlette.responses import HTMLResponse, JSONResponse
+from starlette.responses import HTMLResponse, JSONResponse, Response
 from starlette.staticfiles import StaticFiles
 
 from .translations.data import TRANSLATIONS
@@ -90,9 +94,21 @@ class GenerateFileURLResponseModel(BaseModel):
     file_url: str = Field(description="A URL which the file is accessible on.")
 
 
+class GroupItem(BaseModel):
+    name: str
+    slug: str
+
+
 class GroupedTableNamesResponseModel(BaseModel):
     grouped: t.Dict[str, t.List[str]] = Field(default_factory=list)
     ungrouped: t.List[str] = Field(default_factory=list)
+
+
+class GroupedFormsResponseModel(BaseModel):
+    grouped: t.Dict[str, t.List[FormConfigResponseModel]] = Field(
+        default_factory=list
+    )
+    ungrouped: t.List[FormConfigResponseModel] = Field(default_factory=list)
 
 
 @dataclass
@@ -307,6 +323,16 @@ PydanticModel = t.TypeVar("PydanticModel", bound=BaseModel)
 
 
 @dataclass
+class FileResponse:
+    contents: t.Union[io.StringIO, io.BytesIO]
+    file_name: str
+    media_type: str
+
+
+FormResponse: typing_extensions.TypeAlias = t.Union[str, FileResponse, None]
+
+
+@dataclass
 class FormConfig:
     """
     Used to specify forms, which are passed into ``create_admin``.
@@ -326,6 +352,10 @@ class FormConfig:
     :param description:
         An optional description which is shown in the UI to explain to the user
         what the form is for.
+    :param form_group:
+        If specified, forms can be divided into groups in the form
+        menu. This is useful when you have many forms that you
+        can organize into groups for better visibility.
 
     Here's a full example:
 
@@ -350,7 +380,8 @@ class FormConfig:
         config = FormConfig(
             name="My Form",
             pydantic_model=MyModel,
-            endpoint=my_endpoint
+            endpoint=my_endpoint,
+            form_group="Text forms",
         )
 
     """
@@ -361,14 +392,16 @@ class FormConfig:
         pydantic_model: t.Type[PydanticModel],
         endpoint: t.Callable[
             [Request, PydanticModel],
-            t.Union[str, None, t.Coroutine],
+            t.Union[FormResponse, t.Coroutine[None, None, FormResponse]],
         ],
         description: t.Optional[str] = None,
+        form_group: t.Optional[str] = None,
     ):
         self.name = name
         self.pydantic_model = pydantic_model
         self.endpoint = endpoint
         self.description = description
+        self.form_group = form_group
         self.slug = self.name.replace(" ", "-").lower()
 
 
@@ -431,12 +464,17 @@ class AdminRouter(FastAPI):
         allowed_hosts: t.Sequence[str] = [],
         debug: bool = False,
         sidebar_links: t.Dict[str, str] = {},
+        mfa_providers: t.Optional[t.Sequence[MFAProvider]] = None,
     ) -> None:
         super().__init__(
             title=site_name,
             description=f"{site_name} documentation",
             middleware=[
-                Middleware(CSRFMiddleware, allowed_hosts=allowed_hosts)
+                Middleware(
+                    CSRFMiddleware,
+                    allowed_hosts=allowed_hosts,
+                    allow_form_param=True,
+                )
             ],
             debug=debug,
             exception_handlers={500: log_error},
@@ -604,6 +642,14 @@ class AdminRouter(FastAPI):
         )
 
         private_app.add_api_route(
+            path="/forms/grouped/",
+            endpoint=self.get_grouped_forms,  # type: ignore
+            methods=["GET"],
+            response_model=GroupedFormsResponseModel,
+            tags=["Forms"],
+        )
+
+        private_app.add_api_route(
             path="/forms/{form_slug:str}/",
             endpoint=self.get_single_form,  # type: ignore
             methods=["GET"],
@@ -681,6 +727,30 @@ class AdminRouter(FastAPI):
                         )
 
         #######################################################################
+        # MFA
+
+        if mfa_providers:
+            if len(mfa_providers) > 1:
+                raise ValueError(
+                    "Only a single mfa_provider is currently supported."
+                )
+
+            for mfa_provider in mfa_providers:
+                private_app.mount(
+                    path="/mfa-setup/",
+                    # This rate limiting is because some of the forms accept
+                    # a password, and generating recovery codes is somewhat
+                    # expensive, so we want to prevent abuse.
+                    app=RateLimitingMiddleware(
+                        app=mfa_setup(
+                            provider=mfa_provider,
+                            auth_table=self.auth_table,
+                        ),
+                        provider=InMemoryLimitProvider(limit=20, timespan=300),
+                    ),
+                )
+
+        #######################################################################
 
         public_app = FastAPI(
             redoc_url=None,
@@ -692,11 +762,14 @@ class AdminRouter(FastAPI):
 
         if not rate_limit_provider:
             rate_limit_provider = InMemoryLimitProvider(
-                limit=100, timespan=300
+                limit=20,
+                timespan=300,
             )
 
         public_app.mount(
             path="/login/",
+            # This rate limiting is to prevent brute forcing password login,
+            # and MFA codes.
             app=RateLimitingMiddleware(
                 app=session_login(
                     auth_table=self.auth_table,
@@ -705,6 +778,7 @@ class AdminRouter(FastAPI):
                     max_session_expiry=max_session_expiry,
                     redirect_to=None,
                     production=production,
+                    mfa_providers=mfa_providers,
                 ),
                 provider=rate_limit_provider,
             ),
@@ -895,6 +969,34 @@ class AdminRouter(FastAPI):
             for form in self.forms
         ]
 
+    def get_grouped_forms(self) -> GroupedFormsResponseModel:
+        """
+        Returns a list of custom forms registered with the admin, grouped using
+        `form_group`.
+        """
+        response = GroupedFormsResponseModel()
+        group_names = sorted(
+            {
+                v.form_group
+                for _, v in self.form_config_map.items()
+                if v.form_group
+            }
+        )
+        response.grouped = {i: [] for i in group_names}
+        for _, form_config in self.form_config_map.items():
+            form_group = form_config.form_group
+            form_config_response = FormConfigResponseModel(
+                name=form_config.name,
+                slug=form_config.slug,
+                description=form_config.description,
+            )
+            if form_group is None:
+                response.ungrouped.append(form_config_response)
+            else:
+                response.grouped[form_group].append(form_config_response)
+
+        return response
+
     def get_single_form(self, form_slug: str) -> FormConfigResponseModel:
         """
         Returns the FormConfig for the given form.
@@ -924,10 +1026,11 @@ class AdminRouter(FastAPI):
         Handles posting of custom forms.
         """
         form_config = self.form_config_map.get(form_slug)
-        data = await request.json()
 
         if form_config is None:
             raise HTTPException(status_code=404, detail="No such form found")
+
+        data = await request.json()
 
         try:
             model_instance = form_config.pydantic_model(**data)
@@ -950,6 +1053,18 @@ class AdminRouter(FastAPI):
         except ValueError as exception:
             return JSONResponse(
                 {"custom_form_error": str(exception)}, status_code=422
+            )
+
+        if isinstance(response, FileResponse):
+            headers = {
+                "Content-Disposition": (
+                    f'attachment; filename="{response.file_name}"'
+                )
+            }
+            return Response(
+                response.contents.getvalue(),
+                headers=headers,
+                media_type=response.media_type,
             )
 
         message = (
@@ -1088,6 +1203,7 @@ def create_admin(
     allowed_hosts: t.Sequence[str] = [],
     debug: bool = False,
     sidebar_links: t.Dict[str, str] = {},
+    mfa_providers: t.Optional[t.Sequence[MFAProvider]] = None,
 ):
     """
     :param tables:
@@ -1209,6 +1325,9 @@ def create_admin(
                 },
             )
 
+    :param mfa_providers:
+        Enables Multi-factor Authentication in the login process.
+
     """  # noqa: E501
     auth_table = auth_table or BaseUser
     session_table = session_table or SessionsBase
@@ -1254,4 +1373,5 @@ def create_admin(
         allowed_hosts=allowed_hosts,
         debug=debug,
         sidebar_links=sidebar_links,
+        mfa_providers=mfa_providers,
     )